move tests into a sub module directory
